requirements.txt
streamlit
python-dotenv
langchain
langchain-community
langchain-huggingface
langchain-groq
faiss-cpu
sentence-transformers
PyPDF2
groq
matplotlib